+++ apps/wizard/pages/chart_diff/app.py
@@ -1,5 +1,3 @@
-import difflib
-import json
 from pathlib import Path
 
 import streamlit as st
@@ -9,8 +7,11 @@ from st_pages import add_indentation
 
 from apps.staging_sync.cli import _get_engine_for_env, _modified_chart_ids_by_admin, _validate_env
 from apps.wizard.pages.chart_diff.chart_diff import ChartDiffModified
+from apps.wizard.pages.chart_diff.config_diff import st_show_diff
 from apps.wizard.utils import chart_html, set_states
+from apps.wizard.utils.env import OWID_ENV
 from etl import grapher_model as gm
+from etl.config import DB_HOST
 
 # from apps.wizard import utils as wizard_utils
 
@@ -18,7 +19,7 @@ from etl import grapher_model as gm
 
 CURRENT_DIR = Path(__file__).resolve().parent
 # TODO: unhardcode this
-SOURCE_ENV = "staging-site-streamlit-chart-approval"
+SOURCE_ENV = DB_HOST  # "staging-site-streamlit-chart-approval"
 # TODO: switch to production once we are ready
 TARGET_ENV = "staging-site-master"
 
@@ -40,6 +41,8 @@ add_indentation()
 ########################################
 # FUNCTIONS
 ########################################
+
+
 def get_chart_diffs(source_engine, target_engine):
     with Session(source_engine) as source_session:
         with Session(target_engine) as target_session:
@@ -54,30 +57,6 @@ def get_chart_diffs(source_engine, target_engine):
                 for chart_id in chart_ids
             }
     return chart_diffs
-
-
-def process_diff(diff):
-    processed_diff = []
-    for line in diff:
-        if line.startswith("+"):
-            # Additions in green
-            processed_diff.append(f":green[{line}]")
-        elif line.startswith("-"):
-            # Deletions in red
-            processed_diff.append(f":red[{line}]")
-        else:
-            # Unchanged lines
-            processed_diff.append(line)
-    return processed_diff
-
-
-def st_show_diff(config_1, config_2):
-    config_1 = json.dumps(config_1, indent=4)
-    config_2 = json.dumps(config_2, indent=4)
-    diff = difflib.ndiff(config_1.splitlines(keepends=True), config_2.splitlines(keepends=True))
-    processed_diff = process_diff(diff)
-    with st.container(border=True):
-        st.markdown("<br>".join(processed_diff), unsafe_allow_html=True)
 
 
 def st_show(diff: ChartDiffModified, source_session, target_session=None) -> None:
@@ -195,13 +174,14 @@ def get_engines() -> tuple[Engine, Engine]:
 
 def show_help_text():
     with st.popover("How does this work?"):
+        staging_name = OWID_ENV.name.upper()
         st.markdown(
-            """
-        **Chart diff** is a living page that compares all ongoing charts between `PRODUCTION` and your `STAGING` environment.
+            f"""
+        **Chart diff** is a living page that compares all ongoing charts between `PRODUCTION` and your `{staging_name}` environment.
 
-        It lists all those charts that have been modified in the `STAGING` environment.
+        It lists all those charts that have been modified in the `{staging_name}` environment.
 
-        If you want any of the modified charts in `STAGING` to be migrated to `PRODUCTION`, you can approve them by clicking on the toggle button.
+        If you want any of the modified charts in `{staging_name}` to be migrated to `PRODUCTION`, you can approve them by clicking on the toggle button.
         """
         )
 
@@ -247,13 +227,13 @@ def main():
             with Session(target_engine) as target_session:
                 if chart_diffs_modified:
                     st.header("Modified charts")
-                    st.markdown(f"{len(chart_diffs_modified)} charts modified in `{SOURCE_ENV}`")
+                    st.markdown(f"{len(chart_diffs_modified)} charts modified in [{OWID_ENV.name}]({OWID_ENV.site})")
                     for chart_diff in chart_diffs_modified:
                         st_show(chart_diff, source_session, target_session)
 
                 if chart_diffs_new:
                     st.header("New charts")
-                    st.markdown(f"{len(chart_diffs_new)} new charts in `{SOURCE_ENV}`")
+                    st.markdown(f"{len(chart_diffs_new)} new charts in [{OWID_ENV.name}]({OWID_ENV.site})")
                     for chart_diff in chart_diffs_new:
                         st_show(chart_diff, source_session, target_session)
